+++ src/commands/database-instances-list.ts
@@ -3,10 +3,12 @@ import logger = require("../logger");
 import { requirePermissions } from "../requirePermissions";
 import getProjectNumber = require("../getProjectNumber");
 import firedata = require("../gcp/firedata");
+import { warnRealtimeDatabaseEmulated } from "../emulator/commandUtils";
 
 export default new Command("database:instances:list")
   .description("list realtime database instances")
   .before(requirePermissions, ["firebasedatabase.instances.list"])
+  .before(warnRealtimeDatabaseEmulated, false)
   .action(async (options: any) => {
     const projectNumber = await getProjectNumber(options);
     const instances = await firedata.listDatabaseInstances(projectNumber);