+++ src/commands/database-rules-get.ts
@@ -3,6 +3,7 @@ import * as logger from "../logger";
 import * as requireInstance from "../requireInstance";
 import { requirePermissions } from "../requirePermissions";
 import * as metadata from "../database/metadata";
+import { warnRealtimeDatabaseEmulated } from "../emulator/commandUtils";
 
 export default new Command("database:rules:get <rulesetId>")
   .description("get a realtime database ruleset by id")
@@ -12,6 +13,7 @@ export default new Command("database:rules:get <rulesetId>")
   )
   .before(requirePermissions, ["firebasedatabase.instances.get"])
   .before(requireInstance)
+  .before(warnRealtimeDatabaseEmulated, false)
   .action(async (rulesetId: metadata.RulesetId, options: any) => {
     const ruleset = await metadata.getRuleset(options.instance, rulesetId);
     logger.info(`Ruleset ${ruleset.id} was created at ${ruleset.createdAt}`);