+++ src/commands/database-rules-list.ts
@@ -3,6 +3,7 @@ import * as logger from "../logger";
 import * as requireInstance from "../requireInstance";
 import { requirePermissions } from "../requirePermissions";
 import * as metadata from "../database/metadata";
+import { warnRealtimeDatabaseEmulated } from "../emulator/commandUtils";
 
 export default new Command("database:rules:list")
   .description("list realtime database rulesets")
@@ -12,6 +13,7 @@ export default new Command("database:rules:list")
   )
   .before(requirePermissions, ["firebasedatabase.instances.get"])
   .before(requireInstance)
+  .before(warnRealtimeDatabaseEmulated, false)
   .action(async (options: any) => {
     const labeled = await metadata.getRulesetLabels(options.instance);
     const rulesets = await metadata.listAllRulesets(options.instance);